+++ src/routes/chat/router.ts
@@ -1,37 +1,10 @@
 const { Router } = require("express");
 const router = Router();
-import { Server, Socket } from 'socket.io';
 import authentication from "../../utils/auth";
 
 const createMessage = require("./CREATE_Message");
 const getMessages = require("./GET_Messages");
 // const editMessage= require("./EDIT_Message")
-
- 
-
-export const initializeSocketIO = (server: any) => {
-    const io = new Server(server, {
-        cors: {
-            origin: process.env.FRONTEND_BASE_URL
-        }
-    });
-
-    io.on('connection', (socket: Socket) => {
-        console.log('Client connected:', socket.id);
-
-        socket.on('sendMessage', (data) => {
-            console.log("wtdgfzhfjzk");
-            
-            console.log(data);
-            socket.emit("receiveMessage", data)
-            //createMessage(socket, messageContent);
-        });
-
-        // socket.on('editMessage', (editedMessage: { messageId: string, newContent: string }) => {
-        //     editMessage(editedMessage);
-        // });
-    });
-};
 
 router.get('/messages/:id',authentication, getMessages)
 router.post('/',authentication, createMessage)
